Move code to new public config package
